src/client/requirements.txt
awsiotsdk==1.5.11
awscrt==0.11.8